chore(suite-native): purge obsolete native discovery redux
