+++ suite-native/discovery/package.json
@@ -11,21 +11,12 @@
     },
     "dependencies": {
         "@mobily/ts-belt": "^3.13.1",
-        "@reduxjs/toolkit": "2.8.2",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/token-definitions": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-constants": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
-        "@suite-common/wallet-utils": "workspace:*",
-        "@suite-native/accounts": "workspace:*",
-        "@suite-native/analytics": "workspace:*",
         "@suite-native/config": "workspace:*",
-        "@suite-native/device": "workspace:*",
-        "@suite-native/device-mutex": "workspace:*",
         "@suite-native/feature-flags": "workspace:*",
-        "@trezor/connect": "workspace:*",
         "@trezor/utils": "workspace:*",
         "react": "18.2.0",
         "react-redux": "9.2.0"
