--- suite-native/discovery/src/types.ts
@@ -1,3 +1,0 @@
-import { DiscoveryItem } from '@suite-common/wallet-types';
-
-export type DiscoveryDescriptorItem = DiscoveryItem & { descriptor: string };
