+++ suite-native/discovery/tsconfig.json
@@ -12,24 +12,10 @@
             "path": "../../suite-common/wallet-config"
         },
         {
-            "path": "../../suite-common/wallet-constants"
-        },
-        {
             "path": "../../suite-common/wallet-core"
         },
-        {
-            "path": "../../suite-common/wallet-types"
-        },
-        {
-            "path": "../../suite-common/wallet-utils"
-        },
-        { "path": "../accounts" },
-        { "path": "../analytics" },
         { "path": "../config" },
-        { "path": "../device" },
-        { "path": "../device-mutex" },
         { "path": "../feature-flags" },
-        { "path": "../../packages/connect" },
         { "path": "../../packages/utils" }
     ]
 }
