+++ suite-native/state/src/reducers.ts
@@ -28,7 +28,6 @@ import {
 import { prepareWalletConnectReducer } from '@suite-common/walletconnect/src/walletConnectReducer';
 import { bannerFlagsPersistWhitelist, bannerFlagsReducer } from '@suite-native/banner-flags';
 import { deviceAuthorizationReducer } from '@suite-native/device-authorization';
-import { discoveryConfigReducer } from '@suite-native/discovery';
 import { featureFlagsPersistedKeys, featureFlagsReducer } from '@suite-native/feature-flags';
 import { nativeFirmwareReducer } from '@suite-native/firmware';
 import { graphPersistTransform, graphReducer } from '@suite-native/graph';
@@ -38,14 +37,12 @@ import { appSettingsPersistWhitelist, appSettingsReducer } from '@suite-native/s
 import {
     deriveAccountTypeFromPaymentType,
     devicePersistTransform,
-    discoveryStopPersistTransform,
     initMmkvStorage,
     migrateAccountBnbToBsc,
     migrateAccountLabel,
     migrateAccountsDeprecateNetworks,
     migrateDeviceState,
-    migrateDiscoveryDeprecateNetworks,
-    migrateEnabledDiscoveryNetworkSymbols,
+    migrateDiscoveryConfigToWalletSettings,
     migrateTransactionsBnbToBsc,
     migrateTransactionsDeprecateNetworks,
     preparePersistReducer,
@@ -119,12 +116,17 @@ export const prepareRootReducers = async () => {
                 key: 'discoveryConfig',
                 storage: await initMmkvStorage(),
             })) as any;
-            if (appSettings && discoveryConfig)
+            if (appSettings && discoveryConfig) {
+                const enabledNetworks = migrateDiscoveryConfigToWalletSettings(
+                    discoveryConfig.enabledDiscoveryNetworkSymbols,
+                );
+
                 return {
                     localCurrency: appSettings.fiatCurrencyCode,
-                    enabledNetworks: discoveryConfig.enabledDiscoveryNetworkSymbols,
+                    enabledNetworks,
                     bitcoinAmountUnit: appSettings.bitcoinUnits,
                 };
+            }
         },
     });
 
@@ -200,45 +202,6 @@ export const prepareRootReducers = async () => {
         },
     });
 
-    const discoveryConfigPersistedReducer = await preparePersistReducer({
-        reducer: discoveryConfigReducer,
-        persistedKeys: [],
-        key: 'discoveryConfig',
-        version: 3,
-        migrations: {
-            2: (oldState: any) => {
-                if (!oldState.enabledDiscoveryNetworkSymbols) return oldState;
-
-                const { enabledDiscoveryNetworkSymbols } = oldState;
-                const migrateNetworkSymbols = migrateEnabledDiscoveryNetworkSymbols(
-                    enabledDiscoveryNetworkSymbols,
-                );
-                const migratedState = {
-                    ...oldState,
-                    enabledDiscoveryNetworkSymbols: migrateNetworkSymbols,
-                };
-
-                return migratedState;
-            },
-            3: (oldState: any) => {
-                if (!oldState.enabledDiscoveryNetworkSymbols) return oldState;
-
-                const { enabledDiscoveryNetworkSymbols } = oldState;
-                const migratedNetworkSymbols = migrateDiscoveryDeprecateNetworks(
-                    enabledDiscoveryNetworkSymbols,
-                );
-                const migratedState = {
-                    ...oldState,
-                    enabledDiscoveryNetworkSymbols: migratedNetworkSymbols,
-                };
-
-                return migratedState;
-            },
-        },
-        transforms: [discoveryStopPersistTransform],
-        // kept for backward compatibility, but not persisted anymore
-    });
-
     const featureFlagsPersistedReducer = await preparePersistReducer({
         reducer: featureFlagsReducer,
         persistedKeys: featureFlagsPersistedKeys,
@@ -275,7 +238,6 @@ export const prepareRootReducers = async () => {
             nativeFirmware: nativeFirmwareReducer,
             logs: logsSlice.reducer,
             notifications: notificationsReducer,
-            discoveryConfig: discoveryConfigPersistedReducer,
             messageSystem: messageSystemPersistedReducer,
             tokenDefinitions: tokenDefinitionsReducer,
             connectPopup: connectPopupReducer,
